Refactoring, auto format
